Setup base project template
When we create a new project, we want a standard setup that has a bunch
of our defaults already in place to allow us to get to the actual core
of the thing we're building quicker.

This is the Ruby starter template that setups up Docker (based on a
Heroku Ruby image), the usual array of `bin` scripts for managing the
environment and Rubocop for keeping the style consistent.