Move web client to subfolder
This project is going to use a monorepo style. Separating the web client
into its own subfolder creates the space at the top level for other
folders such as 'functions' that will define the backend.

This monorepo ensures that everything stays in one place and can easily
be accessed. While also ensuring that the version of the front-end that
matches the back-end will be historically accessible.