Experiment with using AWS Lambda functions (#11)
The backend for this version of Openshelf is going to be built on AWS Lambda. I'm experimenting with it and trying to get it working. This is the state of the function where I first got it working.

I had issues with getting the function to be able to access the RDS instance. I hadn't set the inbound security rules on the security group for RDS to include the VPC that the function is in.